ConnectX Kaggle Competition 1
A competition on Apple's App Store dataset.
The dataset is split into training and testing sets. The testing set deliberately excludes the target variable. This ensures that the model's performance evaluation relies solely on its ability to generalize and make accurate predictions based on the features and patterns it has learned from the training data during the preprocessing and preparation stage. This safeguards against the model simply memorizing the target values in the training set, which can lead to poor performance on unseen data.